minor patch
fixing delete user function

+++ env/delete_user.php
@@ -2,14 +2,19 @@
 // Koneksi ke database
 include("config.php");
 
-$user_id = 1; // ganti dengan user_id yang ingin dihapus
-// Buat query untuk menghapus data dari tabel users
-$sql = "DELETE FROM users WHERE user_id='$user_id'";
-$query = mysqli_query($conn, $sql);
+if (isset($_GET['user_id'])) {
+    $user_id = $_GET['user_id'];
+    
+    // Menghapus user berdasarkan user_id
+    $sql = "DELETE FROM users WHERE user_id = '$user_id'";
+    $query = mysqli_query($conn, $sql);
 
-if ($query) {
-    echo '<script>alert("Data berhasil dihapus!"); window.location.href="../pages/admin.php";</script>';
+    if ($query) {
+        echo '<script>alert("Data berhasil dihapus!"); window.location.href="../pages/admin.php";</script>';
+    } else {
+        echo '<script>alert("Terjadi kesalahan saat menghapus data."); window.location.href="../pages/admin.php";</script>';
+    }
 } else {
-    echo '<script>alert("Terjadi kesalahan saat menghapus data."); window.location.href="../pages/admin.php";</script>';
+    echo '<script>alert("User dengan ID tersebut tidak ditemukan."); window.location.href="../pages/admin.php";</script>';
 }
 ?>
